requirements.txt
python-telegram-bot
pandas
numpy
scikit-learn
openai